api/requirements.txt
# Django
django==4.1.7
django-environ~=0.8.1
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-dotenv==1.4.2
django_model_utils-4.3.1

# Scrapy
Scrapy==2.8.0
scrapy-djangoitem==1.1.1